+++ src/yb/consensus/log-test.cc
@@ -661,7 +661,7 @@ TEST_F(LogTest, TestGCWithLogRunning) {
   // Check that we get a NotFound if we try to read before the GCed point.
   {
     ReplicateMsgs repls;
-    long starting_op_segment_seq_num;
+    uint32_t starting_op_segment_seq_num;
     yb::SchemaPB schema;
     uint32_t schema_version;
     Status s = log_->GetLogReader()->ReadReplicatesInRange(
@@ -1058,7 +1058,7 @@ TEST_F(LogTest, TestReadLogWithReplacedReplicates) {
     for (int random_read = 0; random_read < kNumRandomReads; random_read++) {
       auto start_index = RandomUniformInt<int64_t>(gc_index, max_repl_index - 1);
       auto end_index = RandomUniformInt<int64_t>(start_index, max_repl_index);
-      long starting_op_segment_seq_num;
+      uint32_t starting_op_segment_seq_num;
       yb::SchemaPB schema;
       uint32_t schema_version;
       {
@@ -1134,7 +1134,7 @@ TEST_F(LogTest, TestReadReplicatesHighIndex) {
 
   auto* reader = log_->GetLogReader();
   ReplicateMsgs repls;
-  long starting_op_segment_seq_num;
+  uint32_t starting_op_segment_seq_num;
   yb::SchemaPB schema;
   uint32_t schema_version;
   ASSERT_OK(reader->ReadReplicatesInRange(first_log_index, first_log_index + kSequenceLength - 1,
